Move render collection with dividers to a component, fix usages
