+++ src/components/CollectionWithDividers/index.tsx
@@ -1,0 +1,2 @@
+export * from "./CollectionWithDividers";
+export { default } from "./CollectionWithDividers";
