+++ src/misc.tsx
@@ -1,7 +1,4 @@
-import Divider from "@material-ui/core/Divider";
-import initial from "lodash-es/initial";
 import moment from "moment-timezone";
-import React from "react";
 import { MutationFunction, MutationResult } from "react-apollo";
 import { defineMessages, IntlShape } from "react-intl";
 import urlJoin from "url-join";
@@ -37,47 +34,6 @@ export type RequireOnlyOne<T, Keys extends keyof T = keyof T> = Pick<
     [K in Keys]-?: Required<Pick<T, K>> &
       Partial<Record<Exclude<Keys, K>, undefined>>;
   }[Keys];
-
-const defaultRenderDivider = () => <Divider />;
-
-export function renderCollectionWithDividers<T>({
-  collection,
-  renderItem,
-  renderEmpty,
-  renderDivider
-}: {
-  collection: T[];
-  renderItem: (
-    item: T | undefined,
-    index: number | undefined,
-    collection: T[]
-  ) => any;
-  renderDivider?: () => React.ReactNode;
-  renderEmpty?: (collection: T[]) => any;
-}) {
-  const hasNoItemsAndPlaceholder = !renderEmpty && !collection.length;
-
-  if (hasNoItemsAndPlaceholder) {
-    return null;
-  }
-
-  if (!collection.length) {
-    return !!renderEmpty ? renderEmpty(collection) : null;
-  }
-
-  const renderDividerFunction = renderDivider || defaultRenderDivider;
-
-  return initial(
-    collection.reduce(
-      (result, item, index) => [
-        ...result,
-        renderItem(item, index, collection),
-        renderDividerFunction()
-      ],
-      []
-    )
-  );
-}
 
 export function renderCollection<T>(
   collection: T[],
